Added HKCR for current user
