+++ NidhoggClient/NidhoggClient.cpp
@@ -1,97 +1,5 @@
 #include "pch.h"
-#include <sstream>
-#include <fstream>
-
-#include "Nidhogg.h"
-
-enum class Options {
-	Unknown,
-	Add, Remove, Clear, Hide, Unhide, Elevate, Signature, Query, Patch, InjectShellcode, InjectDll, DumpCredentials,
-	ExecuteScript
-};
-
-#define PRINT_ASCII_ART
-
-#ifdef PRINT_ASCII_ART
-constexpr const char* ASCII_ART = R"(                  
-                                 8                           
-                               38                            
-                              988                            
-                        90  79888  3                         
-                       880 8998880 88                        
-                       88899998888088                        
-                      7809999999888086                       
-                     50899999999888888     0                 
-                8     09999999999888888     8                
-               83      999999999998880      08               
-              08      0899  99999 8880      880              
-              88     02  9999999990  3488   488              
-              888  88331   0999992  286     880              
-             4888     84    99999  90     28887              
-              8880      0   22032  8       880               
-               888      5   22232  3     8888                
-               88888         233        08808                
-                  9888        2       8880 9                 
-                 988388888        088888888                  
-                  88      0883 8881     48                   
-                            5888                             
-                              8                                                    
-)";
-#else
-constexpr const char* ASCII_ART = "";
-#endif
-
-void PrintUsage() {
-	std::cout << "[ * ] Possible usage:" << std::endl;
-	std::cout << "\tNidhoggClient.exe process [add | remove | clear | hide | unhide | elevate | signature | query ] [pid] [signer type] [signature signer]" << std::endl;
-	std::cout << "\tNidhoggClient.exe thread [add | remove | clear | hide | unhide | query ] [tid]" << std::endl;
-	std::cout << "\tNidhoggClient.exe module [hide] [pid] [module path]" << std::endl;
-	std::cout << "\tNidhoggClient.exe driver [hide | unhide] [driver path]" << std::endl;
-	std::cout << "\tNidhoggClient.exe file [add | remove | clear | query] [path]" << std::endl;
-	std::cout << "\tNidhoggClient.exe reg [add | remove | clear | hide | unhide | query] [key] [value]" << std::endl;
-	std::cout << "\tNidhoggClient.exe patch [pid] [amsi | etw | module name] [function] [patch comma separated]" << std::endl;
-	std::cout << "\tNidhoggClient.exe shinject [apc | thread] [pid] [shellcode file] [parameter 1] [parameter 2] [parameter 3]" << std::endl;
-	std::cout << "\tNidhoggClient.exe dllinject [apc | thread] [pid] [dll path]" << std::endl;
-	std::cout << "\tNidhoggClient.exe callbacks [query | remove | restore] [callback type] [callback address]" << std::endl;
-	std::cout << "\tNidhoggClient.exe etwti [enable | disable]" << std::endl;
-	std::cout << "\tNidhoggClient.exe dump_creds" << std::endl;
-	std::cout << "\tNidhoggClient.exe port [hide | unhide | query | clear] [port number] [tcp/udp] [remote/local]" << std::endl;
-	std::cout << "\tNidhoggClient.exe exec_script [script_file]" << std::endl;
-}
-
-std::vector<byte> ConvertToVector(std::wstring rawPatch) {
-	int b;
-	std::vector<byte> vec;
-	std::wstringstream rawPatchStream(rawPatch);
-	std::wstringstream byteToAdd;
-
-	for (wchar_t i; rawPatchStream >> i; rawPatchStream.good()) {
-		byteToAdd << std::hex << i;
-
-		if (rawPatchStream.peek() == L',') {
-			rawPatchStream.ignore();
-			byteToAdd >> b;
-			vec.push_back(b);
-			byteToAdd.clear();
-		}
-	}
-	byteToAdd >> b;
-	vec.push_back(b);
-
-	return vec;
-}
-
-int ConvertToInt(std::wstring rawString) {
-	std::wstringstream rawPatchStream(rawString);
-	std::wstringstream convertedString;
-
-	for (wchar_t i; rawPatchStream >> i; rawPatchStream.good()) {
-		convertedString << std::hex << i;
-	}
-
-	return _wtoi(convertedString.str().c_str());
-}
-
+#include "NidhoggClient.h"
 
 int wmain(int argc, const wchar_t* argv[]) {
 	std::vector<DWORD> pids;
