+++ NidhoggClient/NidhoggClient.vcxproj
@@ -131,6 +131,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Helper.cpp" />
     <ClCompile Include="Nidhogg.cpp" />
     <ClCompile Include="NidhoggAntiAnalysis.cpp" />
     <ClCompile Include="NidhoggClient.cpp" />
@@ -144,7 +145,9 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Helper.h" />
     <ClInclude Include="Nidhogg.h" />
+    <ClInclude Include="NidhoggClient.h" />
     <ClInclude Include="NidhoggIoctls.h" />
     <ClInclude Include="NidhoggStructs.h" />
     <ClInclude Include="pch.h" />
