+++ NidhoggClient/pch.h
@@ -4,5 +4,8 @@
 #include <Windows.h>
 #include <vector>
 #include <string>
+#include <sstream>
+#include <fstream>
 #include <sddl.h>
+
 #pragma comment(lib, "advapi32.lib")
